backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
python-dotenv==1.0.1
Pillow==10.4.0
djoser==2.2.3
django-filter==24.3
django-cors-headers==4.5.0
psycopg2-binary==2.9.10
pytest-django==4.5.2
pytest==7.1.3
pytest-lazy-fixture==0.6.3
gunicorn
